+++ packages/people-and-teams/focused-task-close-account/src/index.ts
@@ -1,0 +1,9 @@
+export {
+  default as FocusedTaskCloseAccount,
+} from './components/FocusedTaskCloseAccount';
+export {
+  default as DeleteUserContentPreviewScreen,
+} from './components/DeleteUserContentPreviewScreen';
+export {
+  default as DeleteUserOverviewScreen,
+} from './components/DeleteUserOverviewScreen';